fix: size positions against margin and fractional lots, not whole units
Every XAU/USD signal was being floored to zero units, so a backtest over
1433 H1 bars produced 21 crossovers and 0 trades while reporting success.
Two independent modelling errors caused it:

1. `max_position_value_pct` capped NOTIONAL at 20% of equity. That is not
   a 20% cap, it is 0.2x leverage -- less than unleveraged. With gold at
   ~$4,290 and a $10k account the cap allowed $2,000 of notional, i.e.
   zero ounces.
2. `units` was an int, so any size below 1 oz was unrepresentable.

The cap is now on MARGIN (notional / leverage), and sizing is fractional
and quantised to a lot step. `units` widens to float through Proposal,
OrderSpec and the backtester's Trade.

New pinned constants describe the broker instead of assuming it:
`contract_size`, `min_lot_step` and `leverage`. `leverage` defaults to
1.0 because Alpaca does not margin crypto -- PAXG is non-marginable --
so raising it would backtest conditions the broker does not offer.

Sizing moves to `app/risk/sizing.py` as a pure function shared by the
StrategyRunner and the Backtester, with property-based tests asserting
no cap is ever breached and that leverage never shrinks a position.

Effect: a $100 account now takes the same 13 trades a $250k account
takes, and the metrics are equity-invariant (-0.202R expectancy, 0.58
profit factor) as they should be. Worth noting separately: at leverage
1.0 the margin cap binds at every account size, so the strategy only
ever risks 0.33-0.39% per trade rather than the intended 0.5%.

The baseline strategy remains unprofitable; that was never in question.

+++ app/llm/schemas.py
@@ -41,7 +41,7 @@ class ClosedTradeSummary(BaseModel):
     opened_at: datetime
     closed_at: datetime
     side: str
-    units: int
+    units: float
     entry: float
     exit: float
     pnl_usd: float
